@bchocho 2023-11-17
I'm confused by this cherry-pick. Shouldn't we already have it for free because it's in the 1.8 branch? https://github.com/aptos-labs/aptos-core/commits/aptos-release-v1.8